Moved developer.arm.com functionality into separate file generate-vectors.py with common functions in generate_common.py
