+++ embedding-generation/generate-chunks.py
@@ -17,21 +17,11 @@ import json
 import os
 import re
 import uuid
-import yaml
 import csv
-import datetime
 
-import boto3
-from botocore.exceptions import NoCredentialsError, ClientError
 from bs4 import BeautifulSoup
 import requests
-from requests.adapters import HTTPAdapter
-from urllib3.util.retry import Retry
-from urllib.parse import parse_qs, urlparse, quote
-from dataclasses import dataclass
-from typing import Any, Dict, List, Optional
-from playwright.async_api import async_playwright
-import asyncio
+from urllib.parse import parse_qs, urlparse
 
 from document_chunking import (
     arm_service_url_to_developer_url,
@@ -45,58 +35,22 @@ from document_chunking import (
     source_to_fetch_url,
 )
 
+from generate_common import (
+    Chunk,
+    createChunk,
+    printChunks,
+    chunkSaveAndTrack,
+    fetch_with_logging,
+    register_source,
+    save_sources_csv,
+    load_existing_sources,
+    get_number_of_sources,
+    ensure_intrinsic_chunks_from_s3,
+    yaml_dir,
+    details_file,
+    http_session
+)
 
-# Create a session with retry logic for resilient HTTP requests
-def create_retry_session(retries=5, backoff_factor=1, status_forcelist=(500, 502, 503, 504)):
-    """Create a requests session with automatic retry on failures."""
-    session = requests.Session()
-    retry = Retry(
-        total=retries,
-        read=retries,
-        connect=retries,
-        backoff_factor=backoff_factor,
-        status_forcelist=status_forcelist,
-        allowed_methods=["HEAD", "GET", "OPTIONS"]
-    )
-    adapter = HTTPAdapter(max_retries=retry)
-    session.mount("http://", adapter)
-    session.mount("https://", adapter)
-    return session
-
-# Global session for all HTTP requests
-http_session = create_retry_session()
-
-
-def ensure_intrinsic_chunks_from_s3(local_folder='intrinsic_chunks',
-                                    s3_bucket='arm-github-copilot-extension',
-                                    s3_prefix='embedding_data/intrinsic_chunks/'):
-    """
-    Ensure the local 'intrinsic_chunks' folder exists and is populated with files from S3.
-    If the folder does not exist, create it and download all files from the S3 prefix.
-    """
-    if not os.path.exists(local_folder):
-        os.makedirs(local_folder, exist_ok=True)
-        print(f"Created local folder: {local_folder}")
-        s3 = boto3.client('s3')
-        try:
-            paginator = s3.get_paginator('list_objects_v2')
-            for page in paginator.paginate(Bucket=s3_bucket, Prefix=s3_prefix):
-                for obj in page.get('Contents', []):
-                    key = obj['Key']
-                    if key.endswith('/'):
-                        continue  # skip folders
-                    filename = os.path.basename(key)
-                    local_path = os.path.join(local_folder, filename)
-                    print(f"Downloading {key} to {local_path}")
-                    s3.download_file(s3_bucket, key, local_path)
-        except NoCredentialsError:
-            print("AWS credentials not found. Please configure them.")
-        except ClientError as e:
-            print(f"S3 ClientError: {e}")
-        except Exception as e:
-            print(f"Unexpected error: {e}")
-    else:
-        print(f"Folder '{local_folder}' already exists. Skipping S3 download.")
 
 '''
 To fix:
@@ -104,8 +58,6 @@ To fix:
 2. Learning Path titles must come from index page...send through function along with Graviton.
 '''
 
-yaml_dir = os.getenv('YAML_OUTPUT_DIR', 'yaml_data')
-details_file = os.getenv('CHUNK_DETAILS_FILE', 'info/chunk_details.csv')
 
 chunk_index = 1
 
@@ -116,157 +68,13 @@ cross_platform_lps_dont_duplicate = []
 # multi-megabyte HTML document for every source row.
 ecosystem_dashboard_entries = None
 
-# Global tracking for vector-db-sources.csv
-# Set of URLs already in the CSV (for deduplication)
-known_source_urls = set()
-# List of all source entries (including existing and new)
-# Each entry is a dict: {site_name, license_type, display_name, url, keywords}
-all_sources = []
 
 # Increase the file size limit, which defaults to '131,072'
 csv.field_size_limit(10**9) #1,000,000,000 (1 billion), smaller than 64-bit space but avoids 'python overflowerror'
 
 
-def load_existing_sources(csv_file):
-    """
-    Load existing sources from vector-db-sources.csv into memory.
-    Populates known_source_urls set and all_sources list.
-    """
-    global known_source_urls, all_sources
-    known_source_urls = set()
-    all_sources = []
-    
-    if not os.path.exists(csv_file):
-        print(f"Sources file '{csv_file}' does not exist. Starting fresh.")
-        return
-    
-    with open(csv_file, 'r', newline='', encoding='utf-8') as file:
-        reader = csv.DictReader(file)
-        for row in reader:
-            url = row.get('URL', '').strip()
-            if url:
-                known_source_urls.add(url)
-                all_sources.append({
-                    'site_name': row.get('Site Name', ''),
-                    'license_type': row.get('License Type', ''),
-                    'display_name': row.get('Display Name', ''),
-                    'url': url,
-                    'keywords': row.get('Keywords', '')
-                })
-    
-    print(f"Loaded {len(all_sources)} existing sources from '{csv_file}'")
 
 
-def register_source(site_name, license_type, display_name, url, keywords):
-    """
-    Register a new source URL. If the URL already exists, skip it.
-    Returns True if the source was added, False if it was a duplicate.
-    """
-    global known_source_urls, all_sources
-    
-    # Normalize URL for comparison
-    url = url.strip()
-    
-    if url in known_source_urls:
-        return False
-    
-    known_source_urls.add(url)
-    source_entry = {
-        'site_name': site_name,
-        'license_type': license_type,
-        'display_name': display_name,
-        'url': url,
-        'keywords': keywords if isinstance(keywords, str) else '; '.join(keywords)
-    }
-
-    # Keep discovered sources grouped with their existing site section instead of
-    # appending them to the very end of the CSV and fragmenting that block.
-    insert_at = None
-    for index, existing_source in enumerate(all_sources):
-        if existing_source.get('site_name') == site_name:
-            insert_at = index + 1
-
-    if insert_at is None:
-        all_sources.append(source_entry)
-    else:
-        all_sources.insert(insert_at, source_entry)
-
-    print(f"[NEW SOURCE] {display_name}: {url}")
-    return True
-
-
-def save_sources_csv(csv_file):
-    """
-    Write all sources (existing + new) to vector-db-sources.csv.
-    """
-    with open(csv_file, 'w', newline='', encoding='utf-8') as file:
-        writer = csv.writer(file)
-        writer.writerow(['Site Name', 'License Type', 'Display Name', 'URL', 'Keywords'])
-        for source in all_sources:
-            writer.writerow([
-                source['site_name'],
-                source['license_type'],
-                source['display_name'],
-                source['url'],
-                source['keywords']
-            ])
-    
-    print(f"Saved {len(all_sources)} sources to '{csv_file}'")
-
-class Chunk:
-    def __init__(
-        self,
-        title,
-        url,
-        uuid,
-        keywords,
-        content,
-        heading="",
-        heading_path=None,
-        doc_type="",
-        product="",
-        version="",
-        resolved_url="",
-        content_type="",
-    ):
-        self.title = title
-        self.url = url
-        self.uuid = uuid
-        self.content = content
-        self.heading = heading
-        self.heading_path = heading_path or []
-        self.doc_type = doc_type
-        self.product = product
-        self.version = version
-        self.resolved_url = resolved_url
-        self.content_type = content_type
-
-        # Translate keyword list into comma-separated string, and add similar words to keywords.
-        self.keywords = self.formatKeywords(keywords)
-
-    def formatKeywords(self, keywords):
-        """Format keywords list into a lowercase, comma-separated string."""
-        return ', '.join(k.strip() for k in keywords).lower()
-
-    # Used to dump into a yaml file without difficulty
-    def toDict(self):
-        return {
-            'title': self.title,
-            'url': self.url,
-            'uuid': self.uuid,
-            'keywords': self.keywords,
-            'content': self.content,
-            'heading': self.heading,
-            'heading_path': self.heading_path,
-            'doc_type': self.doc_type,
-            'product': self.product,
-            'version': self.version,
-            'resolved_url': self.resolved_url,
-            'content_type': self.content_type,
-        }
-
-    def __repr__(self):
-        return f"Chunk(title={self.title}, url={self.url}, uuid={self.uuid}, heading={self.heading})"
 
 def build_ecosystem_dashboard_entries():
     """Load and cache package-level snippets from the ecosystem dashboard."""
@@ -529,235 +337,6 @@ def createIntrinsicsDatabaseChunks():
         <sudocode>
     '''
 
-
-@dataclass
-class CapturedSearchRequest:
-    url: str
-    method: str
-    headers: Dict[str, str]
-    post_data: Optional[str]
-    response_json: Dict[str, Any]
-
-async def capture_DeveloperArmComSearch(page_url: str) -> CapturedSearchRequest:
-    apicount = 0
-    def is_search_response(resp) -> bool:
-        nonlocal apicount
-        # print("Testing url "+str(resp.request.method.upper())+" "+str(resp.url))
-        if "coveo.com/rest/search/v2" in resp.url and "querySuggest" not in resp.url:
-            apicount += 1
-            return (
-                resp.request.method.upper() in {"GET", "POST"}
-                and resp.status == 200
-                and apicount == 2
-            )
-        else:
-            return False
-
-    async with async_playwright() as p:
-        browser = await p.chromium.launch(headless=True)
-        page = await browser.new_page()
-
-        async with page.expect_response(is_search_response, timeout=30_000) as response_info:
-            await page.goto(page_url, wait_until="domcontentloaded")
-
-        response = await response_info.value
-        data = await response.json()
-        await browser.close()
-
-        if not(isinstance(data, dict) and "results" in data):
-            raise RuntimeError("No search API response was captured. ")
-        else:
-            return CapturedSearchRequest(
-                url=response.url,
-                method=response.request.method,
-                headers=dict(response.request.headers),
-                post_data=response.request.post_data,
-                response_json=data,
-            )
-
-def replay_DeveloperArmComSearch(
-    captured: CapturedSearchRequest,
-    query: str,
-    first_result: int = 0,
-    number_of_results: int = 48,
-) -> Dict[str, Any]:
-
-    def _merge_headers(base_headers: Dict[str, str]) -> Dict[str, str]:
-        keep = {}
-        drop = {"host", "content-length", "accept-encoding", "connection", "origin", "referer"}
-        for k, v in base_headers.items():
-            if k.lower() not in drop:
-                keep[k] = v
-        keep.setdefault("accept", "application/json, text/plain, */*")
-        keep.setdefault("content-type", "application/json")
-        keep.setdefault("user-agent", "Mozilla/5.0")
-        return keep
-
-    if not captured.post_data:
-        raise RuntimeError("Captured request had no POST body to replay.")
-
-    body = json.loads(captured.post_data)
-    body["q"] = query
-    body["firstResult"] = first_result
-    body["numberOfResults"] = number_of_results
-    headers = _merge_headers(captured.headers)
-
-    r = requests.post(captured.url, headers=headers, json=body, timeout=60)
-    r.raise_for_status()
-    return r.json()
-
-def getDeveloperArmComSearchResults(searchterm: str, searchurl: str, maxitems: int = 20000):
-
-    def extract_result(item: Dict[str, Any]) -> Dict[str, Any]:
-        return {
-            "title": item.get("title") or item.get("raw", {}).get("title"),
-            "url": item.get("clickUri") or item.get("uri") or item.get("url"),
-            "type": item.get("raw", {}).get("navigationhierarchiescontenttype"),
-            "author": item.get("raw", {}).get("author") or item.get("raw", {}).get("sysauthor"),
-            "products": item.get("raw", {}).get("navigationhierarchiesproducts"),
-            "objecttype": item.get("raw", {}).get("objecttype"),
-            "keywords": item.get("raw", {}).get("navigationhierarchiestopics")
-        }
-
-    print('Searching developer.arm.com for "'+searchterm+'"')
-    captured = asyncio.run(capture_DeveloperArmComSearch(searchurl))
-
-    all_rows = []
-    finished = False
-    page_size = 48
-    start = 0
-    while (len(all_rows) < maxitems) and not finished:
-        payload = replay_DeveloperArmComSearch(
-            captured,
-            query=searchterm,
-            first_result=start,
-            number_of_results=page_size,
-        )
-
-        items = [extract_result(x) for x in payload["results"]]
-        all_rows.extend(items)
-        finished = len(payload["results"]) < page_size
-        start += page_size
-    print("Found "+str(len(all_rows))+" results")
-    return all_rows
-
-def processDeveloperArmCom(url, title, type, keywords, emit_chunks=True):
-
-    def chunkizeLearningPath(url, title, keywords):
-        if not emit_chunks:
-            return
-
-        response = fetch_with_logging(url)
-        if response is None:
-            return
-        parsed_document = parse_document_content(
-            source_url=url,
-            resolved_url=url,
-            response_content=response.content,
-            content_type=response.headers.get("content-type", "text/html"),
-            fallback_title=title,
-        )
-        chunk_payloads = chunk_parsed_document(
-            parsed_document,
-            doc_type=type,
-            keywords=keywords,
-        )
-
-        # 5) Create chunks for each snippet by adding metadata
-        for payload in chunk_payloads:
-            chunk = createChunk(
-                payload["content"],
-                url,
-                keywords,
-                payload["title"],
-                heading=payload["heading"],
-                heading_path=payload["heading_path"],
-                doc_type=payload["doc_type"],
-                product=payload["product"],
-                version=payload["version"],
-                resolved_url=payload["resolved_url"],
-                content_type=payload["content_type"],
-            )
-            chunkSaveAndTrack(url,chunk)
-
-
-    response = http_session.get(url, timeout=60)
-    soup = BeautifulSoup(response.text, 'html.parser')
-
-    itemtitle = 'Arm '+type+' - '+(blogtitle.get_text() if (blogtitle := soup.find(id='blog-title')) else title)
-    itemdate = blogdate.get_text() if (blogdate := soup.find(id='blog-date')) else ''
-
-    # Register this learning path as a source
-    register_source(
-        site_name='Arm Developer',
-        license_type='Copyright Arm',
-        display_name=itemtitle,
-        url=url,
-        keywords=keywords
-    )
-    chunkizeLearningPath(url,itemtitle,keywords)
-
-def item_is_relevant(item) -> bool:
-    match item["type"]:
-        case "Guide":
-            return item["title"] in {
-                    "What is SME/SME2?",
-                    "Overview of SME",
-                    "Assembly code",
-                    "Streaming SVE",
-                    "Load and Store",
-                    "Z registers",
-                    "Real world examples",
-                    "ZA storage",
-                    "Predication"
-            }
-
-        case "Programmer's Guide":
-            for pattern in {
-                r"/SME-Overview/",
-                r"/CME",
-                r"/matmul-fp32",
-                r"/lut-gemv-rm-int8",
-                r"/matmul-int8",
-                r"/gemv-cm-int8.+/",
-                r"/109246/.*/Introduction(\?|/The.+/)",
-                r"/Introduction-to-CME",
-                r"/Toolchains-and-model-support/(?!Quick-start)",
-                r"/Memory-access.(?!Implications)",
-                r"/Performance-monitoring",
-                r"/Matrix-Multiply-Unit"
-            }:
-                if re.search(pattern, item["url"]):
-                    return True
-            return False
-
-        case "Blog Post":
-            if item["author"] in {"Zenon_Xiu","KhalidS"} and item["title"][0:4] == "Part" and "SME" in item["title"]:
-                return True
-            if item["author"] == "mweidmann" and item["title"][0:41] == "Introducing the Scalable Matrix Extension":
-                return True
-            return False
-
-def createDeveloperArmComChunks(emit_chunks=True):
-    search_base = "https://developer.arm.com/search#numberOfResults=48&f-navigationhierarchiescontenttype="
-    content_types = [
-        "Blog Post",
-        "Guide",
-        "Programmer's Guide"
-    ]
-
-    search_url = search_base+",".join([quote(x) for x in content_types])+"&q="
-    for searchterm in ["SME"]:
-        pages = getDeveloperArmComSearchResults(searchterm, search_url+searchterm)
-        for page in pages:
-            if item_is_relevant(page):
-                keywords =  list(set( [searchterm] +
-                                    [key for key_list in (page["keywords"] or []) for key in key_list.split(sep="|")] +
-                                    [key for key_list in (page["products"] or []) for key in key_list.split(sep="|")[2:]]))
-                processDeveloperArmCom(page["url"], page["title"], page["type"], keywords, emit_chunks=emit_chunks)
-
-
-
 def processLearningPath(url, type, emit_chunks=True):
     github_raw_link = "https://raw.githubusercontent.com/ArmDeveloperEcosystem/arm-learning-paths/refs/heads/production/content"
     site_link = "https://learn.arm.com"
@@ -1004,31 +583,6 @@ def URLIsValidCheck(url):
         return False
 
 
-def fetch_with_logging(url):
-    try:
-        response = http_session.get(url, timeout=60)
-        response.raise_for_status()
-        return response
-    except requests.exceptions.HTTPError as http_err:
-        print(f"HTTP error occurred: {http_err}")
-        with open('info/errors.csv', 'a', newline='') as csvfile:
-            csv_writer = csv.writer(csvfile)
-            csv_writer.writerow([url, str(http_err)])
-        return None
-    except Exception as err:
-        print(f"Other error occurred: {err}")
-        with open('info/errors.csv', 'a', newline='') as csvfile:
-            csv_writer = csv.writer(csvfile)
-            csv_writer.writerow([url, str(err)])
-        return None
-    except Exception as err:
-        print(f"Other error occurred: {err}")
-        with open('info/errors.csv', 'a', newline='') as csvfile:
-            csv_writer = csv.writer(csvfile)
-            csv_writer.writerow([url,str(err)])
-        return False
-
-
 def obtainMarkdownContentFromGitHubMDFile(gh_url):
     response = http_session.get(gh_url, timeout=60)
     response.raise_for_status()  # Ensure we got a valid response
@@ -1057,48 +611,6 @@ def obtainTextSnippets__Markdown(content, min_words=300, max_words=500, min_fina
         overlap_tokens=max(0, min_final_words // 4),
     )
     return [chunk["content"] for chunk in chunks]
-
-
-def createChunk(
-    text_snippet,
-    WEBSITE_url,
-    keywords,
-    title,
-    heading="",
-    heading_path=None,
-    doc_type="",
-    product="",
-    version="",
-    resolved_url="",
-    content_type="",
-):
-    chunk = Chunk(
-        title        = title,
-        url          = WEBSITE_url,
-        uuid         = str(uuid.uuid4()),
-        keywords     = keywords,
-        content      = text_snippet,
-        heading      = heading,
-        heading_path = heading_path or [],
-        doc_type     = doc_type,
-        product      = product,
-        version      = version,
-        resolved_url = resolved_url,
-        content_type = content_type,
-    )
-
-    return chunk
-
-
-def printChunks(chunks):
-    for chunk_dict in chunks:
-        print('='*100)
-        print("Title:", chunk_dict['title'])
-        print("Keywords:", chunk_dict['keywords'])
-        print("URL:", chunk_dict['url'])
-        print("Unique ID:", chunk_dict['uuid'])
-        print("Content:", chunk_dict['content'])
-        print('='*100)
 
 
 def parse_keywords(keywords_value, title=""):
@@ -1154,14 +666,12 @@ def _arm_topic_links(topic):
         links.extend(_arm_topic_links(child))
     return links
 
-
 def _arm_metadata_keywords(root_data, keywords_value, source_name):
     keywords = parse_keywords(keywords_value, source_name)
     for value in root_data.get("keywords", []) + root_data.get("products", []):
         if value and value not in keywords:
             keywords.append(value)
     return keywords
-
 
 def create_arm_documentation_chunks(source_url, source_name, doc_type, keywords_value):
     root_response = fetch_with_logging(source_to_fetch_url(source_url))
@@ -1209,65 +719,6 @@ def create_arm_documentation_chunks(source_url, source_name, doc_type, keywords_
     return chunks
 
 
-def chunkSaveAndTrack(url,chunk):
-
-    def addNewRow(current_date,chunk_words,chunk_id):
-        return [url,current_date,chunk_words,'1',chunk_id]
-    
-    def addToExistingRow(row,chunk_words,chunk_id):
-        url = row[0] # same URL
-        date = row[1] # same date
-        words = str(int(row[2]) + chunk_words) # update words
-        chunks = row[3] = str(int(row[3]) + 1) # update number of chunks
-        ids = row[4]+ f", {chunk_id}" # update chunk IDs
-        return [url,date,words,chunks,ids]
-
-
-    def recordChunk():
-        current_date = datetime.date.today().strftime('%Y-%m-%d')
-        chunk_words  = len(chunk.content.split())    
-        chunk_id     = f'chunk_{chunk.uuid}'
-
-        new_rows = []
-
-        with open(details_file, mode='r', newline='', encoding='utf-8') as file:
-            csv_reader = csv.reader(file)
-            try:
-                headers = next(csv_reader)  
-                new_rows.append(headers) # keep in memory
-            except StopIteration:
-                pass
-
-            url_found = False  # Track if the URL is found in any row
-            
-            # Loop through all the rows after the header
-            for row in csv_reader:
-                if row[0] == url:
-                    new_rows.append(addToExistingRow(row, chunk_words, chunk_id))  # Modify and append the row
-                    url_found = True  # Mark that the URL was found
-                else:
-                    new_rows.append(row)  # Append the row without modification
-            
-            # If the URL was not found, append a new row
-            if not url_found:
-                new_rows.append(addNewRow(current_date, chunk_words, chunk_id))
-
-
-        # Overwrite csv with new info
-        with open(details_file, mode='w', newline='') as file:
-            csv_writer = csv.writer(file, delimiter=',')
-            csv_writer.writerows(new_rows) 
-
-    # Save chunk
-    file_name = f"{yaml_dir}/chunk_{chunk.uuid}.yaml"
-    with open(file_name, 'w') as file:
-        yaml.dump(chunk.toDict(), file, default_flow_style=False, sort_keys=False)
-
-    # Record chunk
-    recordChunk()
-    print(f"{file_name} === {chunk.title}")
-
-
 def main():
     skip_discovery = os.getenv("SKIP_DISCOVERY", "").lower() in {"1", "true", "yes"}
 
@@ -1312,10 +763,7 @@ def main():
         # b) Ecosystem Dashboard
         createEcosystemDashboardChunks(emit_chunks=False)
 
-        # c) Developer.Arm.Com
-        createDeveloperArmComChunks(emit_chunks=False)
-
-    # d) Intrinsics
+    # c) Intrinsics
     #createIntrinsicsDatabaseChunks()
 
     # 1) Get URLs and details from CSV
@@ -1334,7 +782,7 @@ def main():
     # Save updated sources CSV with all discovered sources
     save_sources_csv(sources_file)
     print(f"\n=== Source tracking complete ===")
-    print(f"Total sources in {sources_file}: {len(all_sources)}")
+    print(f"Total sources in {sources_file}: {get_number_of_sources()}")
 
 
 if __name__ == "__main__":
